Update ozon_name_value_ids computing.

+++ addons/ozon/models/reports/ozon_report_products_revenue_expenses.py
@@ -43,5 +43,5 @@ class OzonNameValue(models.Model):
 
     ozon_categories_id = fields.Many2one("ozon.categories", string="Категория Ozon")
     name = fields.Char(string="Название")
-    value = fields.Float(string="Товар")
+    value = fields.Integer(string="Товар")
